refactor: change directory from /blocks to /components/block
