+++ src/components/block/Header.tsx
@@ -3,7 +3,7 @@ import { Link as ReactLink, useLocation } from "react-router-dom";
 import { Box, Flex, Image, Spacer, useColorMode } from "@chakra-ui/react";
 
 import Logo from "#/assets/logo.png";
-import SubMenu from "#/blocks/SubMenu";
+import SubMenu from "#/components/block/SubMenu";
 import useDeviceWidth from "#/hooks/useDeviceWidth";
 
 const Header: React.FC = () => {
