+++ src/components/block/Preferences.tsx
@@ -1,7 +1,7 @@
 import { Stack, useColorMode } from "@chakra-ui/react";
 import { useAtom } from "jotai";
 
-import AppOptionSwitch from "#/components/AppOptionSwitch";
+import AppOptionSwitch from "#/components/block/AppOptionSwitch";
 import useDeviceWidth from "#/hooks/useDeviceWidth";
 import {
   reversePlayerInfoAtom,
