+++ src/components/board/ShortcutGuide.tsx
