+++ src/components/common/PlayerScoreButton.tsx
