+++ src/components/config/CompactPlayerTable/index.tsx
@@ -5,12 +5,10 @@ import {
   Box,
   Button,
   Checkbox,
-  HStack,
-  IconButton,
+  CheckboxGroup,
   Input,
   InputGroup,
   InputLeftElement,
-  Select,
   Table,
   TableContainer,
   Tag,
@@ -31,15 +29,9 @@ import {
   type ColumnDef,
   type FilterFn,
 } from "@tanstack/react-table";
-import {
-  ArrowNarrowRight,
-  ChevronLeft,
-  ChevronRight,
-  ChevronsLeft,
-  ChevronsRight,
-  Filter,
-} from "tabler-icons-react";
+import { ArrowNarrowRight, Filter } from "tabler-icons-react";
 
+import Navigation from "#/components/common/TableNavigation";
 import db from "#/utils/db";
 import { GameDBPlayerProps, PlayerDBProps } from "#/utils/types";
 
@@ -54,9 +46,16 @@ const CompactPlayerTable: React.FC<CompactPlayerTableProps> = ({
   playerList,
   gamePlayers,
 }) => {
+  const [rowSelection, setRowSelection] = useState({});
   const [searchText, setSearchText] = useState<string>("");
   const [currentSelectedPlayers, setCurrentSelectPlayers] =
     useState(gamePlayers);
+
+  useEffect(() => {
+    db.games.update(game_id, {
+      players: currentSelectedPlayers,
+    });
+  }, [currentSelectedPlayers]);
 
   const onChangeHandler = async (player: PlayerDBProps) => {
     if (
@@ -86,12 +85,6 @@ const CompactPlayerTable: React.FC<CompactPlayerTableProps> = ({
     }
   };
 
-  useEffect(() => {
-    db.games.update(game_id, {
-      players: currentSelectedPlayers,
-    });
-  }, [currentSelectedPlayers]);
-
   const fuzzyFilter: FilterFn<PlayerDBProps> = (row) => {
     const data = row.original;
     return (
@@ -107,14 +100,7 @@ const CompactPlayerTable: React.FC<CompactPlayerTableProps> = ({
     columnHelper.accessor("id", {
       header: "",
       cell: (info) => {
-        return (
-          <Checkbox
-            isChecked={currentSelectedPlayers
-              .map((gamePlayer) => gamePlayer.id)
-              .includes(info.getValue())}
-            onChange={() => onChangeHandler(info.row.original)}
-          />
-        );
+        return <Checkbox value={info.getValue()} />;
       },
       footer: (info) => info.column.id,
     }),
@@ -202,74 +188,28 @@ const CompactPlayerTable: React.FC<CompactPlayerTableProps> = ({
                 ))}
               </Thead>
               <Tbody>
-                {table.getRowModel().rows.map((row) => {
-                  return (
-                    <Tr key={row.id}>
-                      {row.getVisibleCells().map((cell) => {
-                        return (
-                          <Td key={cell.id}>
-                            {flexRender(
-                              cell.column.columnDef.cell,
-                              cell.getContext()
-                            )}
-                          </Td>
-                        );
-                      })}
-                    </Tr>
-                  );
-                })}
+                <CheckboxGroup defaultValue={gamePlayers.map((p) => p.id)}>
+                  {table.getRowModel().rows.map((row) => {
+                    return (
+                      <Tr key={row.id}>
+                        {row.getVisibleCells().map((cell) => {
+                          return (
+                            <Td key={cell.id}>
+                              {flexRender(
+                                cell.column.columnDef.cell,
+                                cell.getContext()
+                              )}
+                            </Td>
+                          );
+                        })}
+                      </Tr>
+                    );
+                  })}
+                </CheckboxGroup>
               </Tbody>
             </Table>
           </TableContainer>
-          <HStack pt={3}>
-            <IconButton
-              aria-label="最初のページに移動"
-              disabled={!table.getCanPreviousPage()}
-              icon={<ChevronsLeft />}
-              onClick={() => table.setPageIndex(0)}
-              size="xs"
-            />
-            <IconButton
-              aria-label="1ページ戻る"
-              disabled={!table.getCanPreviousPage()}
-              icon={<ChevronLeft />}
-              onClick={() => table.previousPage()}
-              size="xs"
-            />
-            <div>
-              {table.getState().pagination.pageIndex + 1} /{" "}
-              {table.getPageCount()}
-            </div>
-            <IconButton
-              aria-label="1ページ進む"
-              disabled={!table.getCanNextPage()}
-              icon={<ChevronRight />}
-              onClick={() => table.nextPage()}
-              size="xs"
-            />
-            <IconButton
-              aria-label="最後のページに移動"
-              disabled={!table.getCanNextPage()}
-              icon={<ChevronsRight />}
-              onClick={() => table.setPageIndex(table.getPageCount() - 1)}
-              size="xs"
-            />
-            <Box>
-              <Select
-                onChange={(e) => {
-                  table.setPageSize(Number(e.target.value));
-                }}
-                size="sm"
-                value={table.getState().pagination.pageSize}
-              >
-                {[10, 20, 30, 40, 50].map((pageSize) => (
-                  <option key={pageSize} value={pageSize}>
-                    {pageSize}件
-                  </option>
-                ))}
-              </Select>
-            </Box>
-          </HStack>
+          <Navigation table={table} />
           <Box sx={{ pt: 3, textAlign: "right" }}>
             <Button
               as={ReactLink}
